+++ pkg/cel/compiler/env.go
@@ -1,4 +1,4 @@
-package cel
+package compiler
 
 import (
 	"github.com/google/cel-go/cel"
