+++ pkg/cel/compiler/env_test.go
@@ -1,4 +1,4 @@
-package cel
+package compiler
 
 import (
 	"testing"
