+++ pkg/cel/compiler/request.go
@@ -1,0 +1,7 @@
+package compiler
+
+import (
+	apiservercel "k8s.io/apiserver/pkg/admission/plugin/cel"
+)
+
+var RequestType = apiservercel.BuildRequestType()
