+++ pkg/cel/engine/engine.go
@@ -1,0 +1,9 @@
+package engine
+
+import (
+	"context"
+)
+
+type Engine interface {
+	Handle(context.Context, EngineRequest) (EngineResponse, error)
+}
