+++ pkg/cel/engine/provider.go
@@ -23,6 +23,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
+type PProvider[T any] interface {
+	Fetch(context.Context) ([]T, error)
+}
+
 type CompiledValidatingPolicy struct {
 	Actions        sets.Set[admissionregistrationv1.ValidationAction]
 	Policy         policiesv1alpha1.ValidatingPolicy
@@ -36,12 +40,10 @@ type CompiledImageValidatingPolicy struct {
 }
 
 type Provider interface {
-	CompiledValidationPolicies(context.Context) ([]CompiledValidatingPolicy, error)
 	ImageVerificationPolicies(context.Context) ([]CompiledImageValidatingPolicy, error)
 }
 
 type reconcilers struct {
-	*policyReconciler
 	*ivpolpolicyReconciler
 }
 
@@ -57,36 +59,7 @@ func (f ImageValidatingPolProviderFunc) ImageValidatingPolicies(ctx context.Cont
 	return f(ctx)
 }
 
-func NewProvider(compiler policy.Compiler, vpolicies []policiesv1alpha1.ValidatingPolicy, exceptions []*policiesv1alpha1.PolicyException) (VPolProviderFunc, error) {
-	compiled := make([]CompiledValidatingPolicy, 0, len(vpolicies))
-	for _, vp := range vpolicies {
-		var matchedExceptions []*policiesv1alpha1.PolicyException
-		for _, polex := range exceptions {
-			for _, ref := range polex.Spec.PolicyRefs {
-				if ref.Name == vp.GetName() && ref.Kind == vp.GetKind() {
-					matchedExceptions = append(matchedExceptions, polex)
-				}
-			}
-		}
-		policy, err := compiler.CompileValidating(&vp, matchedExceptions)
-		if err != nil {
-			return nil, fmt.Errorf("failed to compile policy %s (%w)", vp.GetName(), err.ToAggregate())
-		}
-		actions := sets.New(vp.Spec.ValidationActions()...)
-		compiled = append(compiled, CompiledValidatingPolicy{
-			Actions:        actions,
-			Policy:         vp,
-			CompiledPolicy: policy,
-		})
-	}
-	provider := func(context.Context) ([]CompiledValidatingPolicy, error) {
-		return compiled, nil
-	}
-	return provider, nil
-}
-
 func NewKubeProvider(
-	compiler policy.Compiler,
 	mgr ctrl.Manager,
 	polexLister policiesv1alpha1listers.PolicyExceptionLister,
 ) (Provider, error) {
@@ -134,17 +107,8 @@ func NewKubeProvider(
 			}
 		},
 	}
-	r := newPolicyReconciler(compiler, mgr.GetClient(), polexLister)
-	err := ctrl.NewControllerManagedBy(mgr).
-		For(&policiesv1alpha1.ValidatingPolicy{}).
-		Watches(&policiesv1alpha1.PolicyException{}, exceptionHandlerFuncs).
-		Complete(r)
-	if err != nil {
-		return nil, fmt.Errorf("failed to construct manager: %w", err)
-	}
-
 	ivpolr := newivPolicyReconciler(mgr.GetClient(), polexLister)
-	err = ctrl.NewControllerManagedBy(mgr).
+	err := ctrl.NewControllerManagedBy(mgr).
 		For(&policiesv1alpha1.ImageValidatingPolicy{}).
 		Watches(&policiesv1alpha1.PolicyException{}, exceptionHandlerFuncs).
 		Complete(ivpolr)
@@ -152,76 +116,7 @@ func NewKubeProvider(
 		return nil, fmt.Errorf("failed to construct manager: %w", err)
 	}
 
-	return reconcilers{r, ivpolr}, nil
-}
-
-type policyReconciler struct {
-	client      client.Client
-	compiler    policy.Compiler
-	lock        *sync.RWMutex
-	policies    map[string]CompiledValidatingPolicy
-	polexLister policiesv1alpha1listers.PolicyExceptionLister
-}
-
-func newPolicyReconciler(
-	compiler policy.Compiler,
-	client client.Client,
-	polexLister policiesv1alpha1listers.PolicyExceptionLister,
-) *policyReconciler {
-	return &policyReconciler{
-		client:      client,
-		compiler:    compiler,
-		lock:        &sync.RWMutex{},
-		policies:    map[string]CompiledValidatingPolicy{},
-		polexLister: polexLister,
-	}
-}
-
-func (r *policyReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	var policy policiesv1alpha1.ValidatingPolicy
-	err := r.client.Get(ctx, req.NamespacedName, &policy)
-	if errors.IsNotFound(err) {
-		r.lock.Lock()
-		defer r.lock.Unlock()
-		delete(r.policies, req.NamespacedName.String())
-		return ctrl.Result{}, nil
-	}
-	if err != nil {
-		return ctrl.Result{}, err
-	}
-
-	if policy.GetStatus().Generated {
-		r.lock.Lock()
-		defer r.lock.Unlock()
-		delete(r.policies, req.NamespacedName.String())
-		return ctrl.Result{}, nil
-	}
-	// get exceptions that match the policy
-	exceptions, err := listExceptions(r.polexLister, policy.GetName(), policy.GetKind())
-	if err != nil {
-		return ctrl.Result{}, err
-	}
-	compiled, errs := r.compiler.CompileValidating(&policy, exceptions)
-	if len(errs) > 0 {
-		fmt.Println(errs)
-		// No need to retry it
-		return ctrl.Result{}, nil
-	}
-	r.lock.Lock()
-	defer r.lock.Unlock()
-	actions := sets.New(policy.Spec.ValidationActions()...)
-	r.policies[req.NamespacedName.String()] = CompiledValidatingPolicy{
-		Actions:        actions,
-		Policy:         policy,
-		CompiledPolicy: compiled,
-	}
-	return ctrl.Result{}, nil
-}
-
-func (r *policyReconciler) CompiledValidationPolicies(ctx context.Context) ([]CompiledValidatingPolicy, error) {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return maps.Values(r.policies), nil
+	return reconcilers{ivpolr}, nil
 }
 
 func listExceptions(polexLister policiesv1alpha1listers.PolicyExceptionLister, policyName, policyKind string) ([]*policiesv1alpha1.PolicyException, error) {
