+++ pkg/cel/policies/vpol/validate.go
@@ -1,18 +1,19 @@
-package policy
+package vpol
 
-import "github.com/kyverno/kyverno/api/policies.kyverno.io/v1alpha1"
+import (
+	"github.com/kyverno/kyverno/api/policies.kyverno.io/v1alpha1"
+	"github.com/kyverno/kyverno/pkg/cel/policies/vpol/compiler"
+)
 
 func Validate(vpol *v1alpha1.ValidatingPolicy) ([]string, error) {
-	compiler := NewCompiler()
-	_, err := compiler.CompileValidating(vpol, nil)
+	compiler := compiler.NewCompiler()
+	_, err := compiler.Compile(vpol, nil)
 	if err == nil {
 		return nil, nil
 	}
-
 	warnings := make([]string, 0, len(err.ToAggregate().Errors()))
 	for _, e := range err.ToAggregate().Errors() {
 		warnings = append(warnings, e.Error())
 	}
-
 	return warnings, err.ToAggregate()
 }
