--- pkg/cel/policy/types.go
@@ -1,10 +1,0 @@
-package policy
-
-import (
-	apiservercel "k8s.io/apiserver/pkg/admission/plugin/cel"
-)
-
-var (
-	NamespaceType = apiservercel.BuildNamespaceType()
-	RequestType   = apiservercel.BuildRequestType()
-)
