+++ pkg/imageverification/evaluator/policy.go
@@ -10,6 +10,7 @@ import (
 	"github.com/google/cel-go/common/types/ref"
 	"github.com/kyverno/kyverno/api/policies.kyverno.io/v1alpha1"
 	policiesv1alpha1 "github.com/kyverno/kyverno/api/policies.kyverno.io/v1alpha1"
+	engine "github.com/kyverno/kyverno/pkg/cel/compiler"
 	"github.com/kyverno/kyverno/pkg/cel/libs/globalcontext"
 	"github.com/kyverno/kyverno/pkg/cel/libs/imagedata"
 	"github.com/kyverno/kyverno/pkg/cel/libs/imageverify"
@@ -43,12 +44,12 @@ type compiledPolicy struct {
 	failurePolicy   admissionregistrationv1.FailurePolicyType
 	matchConditions []cel.Program
 	imageRules      []*match.CompiledMatch
-	verifications   []policy.CompiledValidation
+	verifications   []engine.Validation
 	imageExtractors []*variables.CompiledImageExtractor
 	attestorList    map[string]string
 	attestationList map[string]string
 	creds           *v1alpha1.Credentials
-	exceptions      []policy.CompiledException
+	exceptions      []engine.Exception
 	variables       map[string]cel.Program
 }
 
@@ -79,7 +80,7 @@ func (c *compiledPolicy) Evaluate(ctx context.Context, ictx imagedataloader.Imag
 	}
 
 	data := map[string]any{}
-	vars := lazy.NewMapValue(policy.VariablesType)
+	vars := lazy.NewMapValue(engine.VariablesType)
 	for name, variable := range c.variables {
 		vars.Append(name, func(*lazy.MapValue) ref.Val {
 			out, _, err := variable.ContextEval(ctx, data)
@@ -114,10 +115,10 @@ func (c *compiledPolicy) Evaluate(ctx context.Context, ictx imagedataloader.Imag
 		data[RequestKey] = requestVal.Object
 		data[ObjectKey] = objectVal
 		data[OldObjectKey] = oldObjectVal
-		data[policy.VariablesKey] = vars
-		data[policy.GlobalContextKey] = globalcontext.Context{ContextInterface: context}
-		data[policy.ImageDataKey] = imagedata.Context{ContextInterface: context}
-		data[policy.ResourceKey] = resource.Context{ContextInterface: context}
+		data[engine.VariablesKey] = vars
+		data[engine.GlobalContextKey] = globalcontext.Context{ContextInterface: context}
+		data[engine.ImageDataKey] = imagedata.Context{ContextInterface: context}
+		data[engine.ResourceKey] = resource.Context{ContextInterface: context}
 	} else {
 		data[ObjectKey] = request
 	}
